+++ RRGame/Classes/Scenes/RRGameScene/RRGameScene.h
@@ -28,10 +28,10 @@
 
 
 @interface RRGameScene : CCScene {
-    uint  _numberOfPlayers;
+    NSUInteger  _numberOfPlayers;
 }
 
-+ (id)sceneWithGameMode:(RRGameMode)gameMode numberOfPlayers:(uint)numberOfPlayers firstPlayerColor:(RRPlayerColor)playerColor;
-- (id)initWithGameMode:(RRGameMode)gameMode numberOfPlayers:(uint)numberOfPlayers playerColor:(RRPlayerColor)playerColor;
++ (id)sceneWithGameMode:(RRGameMode)gameMode numberOfPlayers:(NSUInteger)numberOfPlayers firstPlayerColor:(RRPlayerColor)playerColor;
+- (id)initWithGameMode:(RRGameMode)gameMode numberOfPlayers:(NSUInteger)numberOfPlayers playerColor:(RRPlayerColor)playerColor;
 
 @end
